Initial Kotlin-Traverser
The new traverser is based on immutable state-objects for the various subsystems: Mothership, Traverser and (not added yet) Navigation.
This should allow adding erroneous navigation implementations which can be used to assert that a planet has a certain "minimum difficulty".
Because of the missing navigation module the traverser currently only follows a chain of PathSelects starting at the start node (illegal for real planets) and stops once a point without active PathSelect is encountered (PathSelects are only "sent" once which means that cycles are okay).