Added Navigator
Traverser should now be able to use a modified dijkstra for exploration and target-pathing.
The current algorithm produced multiple summaries which seemed identical, this is however (proabably) only because the summary does not include the originally selected path if it is overwritten by a pathSelect-Message. Cases like "force robot against a known bottle" will always trigger this. The current workaround attempts to only propagate  exploration-attempts which align with present pathSelects and if it fails to do so altough a pathSelect is present, selects one attempt to be corrected instead of having all of them be corrected in the same way.

The traverser has successfully completed:
- Bounty
- Vim
- Atom
- PyCharm
- Spyder
- Eclipse
- PrinzenRolle (discovered an error in the planet-design: 0,1,N 1,1,N should cost at least 2)
- Piech
- Anin
- Pizza
- Pico