Add user-based group-lookup
This involves many requests to the gitlab server, to mitigate this caching was also implemented.
By default (configurable in config) the server caches gitlab-results for 30m, if there is a failing lookup (eg. User logs in but is in no project) another timeout is used which defaults to 2m.
(Dev-Server has been configured for 2m default and 30s on lookup-miss)

Closes #38