Implement Planet Testing
This implements the core functionality of #41
In the this version there is no way for the user to start this process via the GUI.
One of the easiest way to get started with the Tester is using the following expression:
LookupPlanet(planet).testWith(DefaultTraverser(planet)).traverseToLogger()
When traversing the same planet multiple times it is recommended to convert the LookupPlanet into a TestPlanet once (.buildTestPlanet()) so that this step is not repeated for every traversal.
The resulting TestTraversal-object implements among other things the Iterable<TestState> interface and can be used in for(...) loops. All intermediate states are returned by default, refer to TestTraversal.traverseToLogger() for an example of status-based logging.